sentiment_analysis_backend/requirements.txt
django
djangorestframework
django-cors-headers
transformers
torch
djangorestframework-simplejwt